+++ test_runner/src/main/kotlin/ftl/reports/JUnitReport.kt
@@ -1,5 +1,6 @@
 package ftl.reports
 
+import ftl.args.IArgs
 import ftl.json.MatrixMap
 import ftl.reports.util.IReport
 import ftl.reports.xml.model.JUnitTestResult
@@ -10,18 +11,18 @@ import ftl.util.Utils.write
 object JUnitReport : IReport {
     override val extension = ".xml"
 
-    private fun write(matrices: MatrixMap, output: String) {
-        val reportPath = reportPath(matrices)
+    private fun write(matrices: MatrixMap, output: String, args: IArgs) {
+        val reportPath = reportPath(matrices, args)
         reportPath.write(output)
     }
 
-    override fun run(matrices: MatrixMap, testSuite: JUnitTestResult?, printToStdout: Boolean) {
+    override fun run(matrices: MatrixMap, testSuite: JUnitTestResult?, printToStdout: Boolean, args: IArgs) {
         val output = testSuite.xmlToString()
 
         if (printToStdout) {
             print(output)
         } else {
-            write(matrices, output)
+            write(matrices, output, args)
         }
     }
 }